Rename to ExampleCrowdControlledDebuff, other adjustments
