+++ ExampleMod/Localization/TranslationsNeeded.txt
@@ -1,3 +1,3 @@
-en-US, 743/743, 100%, missing 0
-ru-RU, 6/743, 1%, missing 737
-zh-Hans, 2/743, 0%, missing 741
+en-US, 745/745, 100%, missing 0
+ru-RU, 6/745, 1%, missing 739
+zh-Hans, 2/745, 0%, missing 743
